refactor(samples): extract shared view models + WopiOptions into WopiHost.Web.Shared (#456)
Addresses the Low-severity dedup item on #456 with the scope the user picked: pure data
types only — Razor components / token minters / Program.cs stay sample-local because they
have legitimate per-sample divergence (auth chrome, OIDC claim extraction, postMessage
handshake) that would only obscure the samples-as-templates property if shared.

New project: `sample/WopiHost.Web.Shared` containing:
- `BrowseViewModel.cs` + `BreadcrumbPart` record
- `ContainerViewModel.cs`
- `FileViewModel.cs`
- `ErrorViewModel.cs`
- `WopiOptions.cs` (keeps the anonymous-sample's `UiCulture` setting; OIDC's Detail.razor
  now resolves `WopiUrlSettings.UiLlcc` from it instead of hard-coding `new CultureInfo("en-US")`)

All five files moved via `git mv` so blame/history is preserved. The OIDC sample's
duplicates are deleted; both samples now `using WopiHost.Web.Shared;` (added to each
`_Imports.razor` and `Program.cs`). OIDC keeps its own `OidcOptions` + `WopiSigningOptions`
in `Models/` since those are OIDC-specific.

Wopi:Web/Models directory deleted (empty). Smoke-test factory updated.
