+++ sample/WopiHost.Web.Oidc/Components/_Imports.razor
@@ -8,3 +8,4 @@
 @using WopiHost.Web.Oidc.Components
 @using WopiHost.Web.Oidc.Components.Layout
 @using WopiHost.Web.Oidc.Models
+@using WopiHost.Web.Shared
