--- sample/WopiHost.Web.Oidc/Models/ContainerViewModel.cs
@@ -1,7 +1,0 @@
-namespace WopiHost.Web.Oidc.Models;
-
-public class ContainerViewModel
-{
-    public required string ContainerId { get; set; }
-    public required string Name { get; set; }
-}
