--- sample/WopiHost.Web.Oidc/Models/ErrorViewModel.cs
@@ -1,7 +1,0 @@
-namespace WopiHost.Web.Oidc.Models;
-
-public class ErrorViewModel
-{
-    public Exception? Exception { get; set; }
-    public bool ShowExceptionDetails { get; set; }
-}
