+++ sample/WopiHost.Web.Shared/ContainerViewModel.cs
@@ -1,0 +1,7 @@
+namespace WopiHost.Web.Shared;
+
+public class ContainerViewModel
+{
+    public required string ContainerId { get; set; }
+    public required string Name { get; set; }
+}
