+++ sample/WopiHost.Web.Shared/WopiHost.Web.Shared.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<Nullable>enable</Nullable>
+		<TargetFramework>net10.0</TargetFramework>
+		<AssemblyName>WopiHost.Web.Shared</AssemblyName>
+		<RootNamespace>WopiHost.Web.Shared</RootNamespace>
+		<IsPackable>false</IsPackable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\..\src\WopiHost.Discovery\WopiHost.Discovery.csproj" />
+	</ItemGroup>
+
+</Project>
