+++ sample/WopiHost.Web/Components/Pages/Browse.razor
@@ -6,7 +6,7 @@
 @using WopiHost.Abstractions
 @using WopiHost.Discovery
 @using WopiHost.Discovery.Enumerations
-@using WopiHost.Web.Models
+@using WopiHost.Web.Shared
 
 @inject IWopiStorageProvider Storage
 @inject IDiscoverer Discoverer
