+++ sample/WopiHost.Web/Components/Pages/Detail.razor
@@ -7,7 +7,7 @@
 @using WopiHost.Discovery
 @using WopiHost.Discovery.Enumerations
 @using WopiHost.Url
-@using WopiHost.Web.Models
+@using WopiHost.Web.Shared
 @using WopiHost.Web.Services
 
 @inject IOptions<WopiOptions> WopiOptionsAccessor
