+++ sample/WopiHost.Web/Components/_Imports.razor
@@ -5,4 +5,4 @@
 @using WopiHost.Web
 @using WopiHost.Web.Components
 @using WopiHost.Web.Components.Layout
-@using WopiHost.Web.Models
+@using WopiHost.Web.Shared
