--- sample/WopiHost.Web/Models/ContainerViewModel.cs
@@ -1,7 +1,0 @@
-namespace WopiHost.Web.Models;
-
-public class ContainerViewModel
-{
-    public required string ContainerId { get; set; }
-    public required string Name { get; set; }
-}
